Allow git@github.com"/" as well as git@github.com":"
When you clone URLs using ghq like so:

```
    ghq get git@github.com:lestrrat/Data-Localize.git
```

It produces git remote origin in the form of

```
    url = ssh://git@github.com/lestrrat/Data-Localize.git
```

Notice the ssh:// scheme (which was addressed in 7eabe9c), and the slash
after the "github.com".

Minilla however expects

```
    [\w\-]+ @github.com:
```

The original regex expects a ":" in between the hostname part and the
username part, where as ghq's URL style uses a "/". This change now
allows both