Ask Git for the top level, rather than inferring from .git/
This fixes an issue where `minil` would try to act on the parent module rather than the submodule when in the submodule’s directory.

When inside a Git submodule, `.git` is a file, not a directory.  By asking Git itself for the top level, we don’t have to worry about such complications.